Update compensation request issue template
This change updates the compensation request issue template to reflect
the sections I used in my request last month at #68. It more clearly
distinguishes between delivered work and works in progress and calls out
a separate section for roles performed.

The goal is that this change is being made in time for most
@bisq-network/contributors who intend to submit compensation requests
for the June 1st–3rd voting round at #70 (several have already
submitted, and so will miss this template change, but can adapt
their existing requests accordingly if they like).